Original-Python-Experiment/requirements.txt
# Core Data Analysis and Scientific Computing
numpy>=1.16.0,<1.22.0
pandas>=1.0.0,<1.6.0
scipy>=1.7.0,<1.10.0

# Financial Data Sources and APIs
yfinance>=0.1.50,<0.2.0
alpha_vantage>=2.3.0,<3.0.0
requests>=2.25.0,<3.0.0

# Technical Analysis and Trading
TA-Lib>=0.4.24,<0.5.0
talib-binary>=0.4.19
backtrader>=1.9.76,<2.0.0
bt>=0.2.9
vectorbt>=0.24.0,<0.25.0
zipline-reloaded>=2.2.0,<3.0.0

# Visualization and Charting
matplotlib>=3.1.0,<3.6.0
plotly>=5.0.0,<6.0.0
seaborn>=0.11.0,<0.13.0
mplfinance>=0.12.0,<0.13.0

# Database connectivity
mysql-connector-python>=8.0.0,<9.0.0
SQLAlchemy>=1.4.0,<2.0.0
PyYAML>=5.1.0,<7.0.0

# Machine Learning and AI Integration
scikit-learn>=1.0.0,<1.2.0
openai>=0.27.0,<1.0.0
anthropic>=0.3.0,<1.0.0

# Risk Management and Portfolio Optimization
pyfolio>=0.9.2
empyrical>=0.5.3
quantlib>=1.26
PyPortfolioOpt>=1.5.0

# Performance and Utility
joblib>=1.0.0,<2.0.0
tqdm>=4.62.0,<5.0.0
cachetools>=4.2.0,<6.0.0

# PHP Development and Testing Requirements
# Note: These are documented for reference, install via composer for PHP components

# PHP Testing Framework (install via composer)
# composer require --dev phpunit/phpunit:^9.6

# PHP Security Framework (install via composer)  
# composer require symfony/security-core:^6.0
# composer require symfony/security-csrf:^6.0
# composer require firebase/php-jwt:^6.0

# PHP HTTP Foundation (install via composer)
# composer require symfony/http-foundation:^6.0

# Development Quality Tools (install via composer)
# composer require --dev phpstan/phpstan
# composer require --dev squizlabs/php_codesniffer
# composer require --dev phpmd/phpmd

# Enhanced Bank Import System Components (PHP 8.4.6+)
# ==================================================
# Architecture: Factory Pattern + Strategy Pattern + Dependency Injection
# 
# Core Components:
# - web_ui/parsers/TransactionParserInterface.php (Parser contract)
# - web_ui/parsers/ParserFactory.php (Factory for parser management)
# - web_ui/parsers/CsvFileReader.php (File I/O handler with DI)
# - web_ui/parsers/CibcParser.php (CIBC bank format parser)
# - web_ui/parsers/MidCapParser.php (Generic CSV format parser)
# - web_ui/bank_import.php (Enhanced UI with validation)
# - tests/unit/ParserTest.php (Comprehensive unit tests - 44 assertions)
#
# Features:
# - SOLID Principles compliance (SRP, OCP, DIP)
# - Auto-detection of CSV formats with fallback suggestions
# - Bank account selection with access validation
# - File upload validation (size, type, security)
# - Parser-specific validation with user-friendly error messages
# - Comprehensive unit test coverage (100% pass rate)
# - PHPDoc documentation with PlantUML diagrams
# - XSS protection and input sanitization
# - Memory-efficient CSV processing
# - Extensible architecture for adding new bank parsers
#
# Testing Framework:
# - PHPUnit 9.6.25 with 7 test methods
# - Mock data for CIBC and MidCap formats
# - Edge case and error condition testing
# - Validation testing for format detection
# - Security testing for input handling
#
# Architecture Documentation:
# - ENHANCED_IMPORT_SYSTEM.md (Comprehensive system documentation)
# - Enhanced_Import_Architecture.puml (PlantUML architecture diagram)
# - Inline PHPDoc with UML class diagrams
#
# Security Features:
# - File type validation (CSV only)
# - File size limits (10MB maximum)
# - Bank account access control per user
# - Input sanitization and XSS protection
# - Upload error handling with descriptive messages
